// src/lib/data/canadaProvinces.generated.ts
export interface ProvincePolygon {
  exterior: [number, number][];
  holes: [number, number][][];
}

export interface ProvinceData {
  name: string;
  iso: string;
  centroidLat: number;
  centroidLon: number;
  polygons: ProvincePolygon[];
}